Revert "client refactor in progress"
This reverts commit a64c2c66e84dbf320021be5b32dbd7e55882d2f5.
