+++ client/examples/scheduler.rs
@@ -3,7 +3,6 @@ use client::{
     execute_task, initialize_client, initialize_scheduler, push_task, setup_payer, setup_program,
     setup_scheduler_account, Config,
 };
-use serde_json;
 
 /// Main entry point for the Solana program client
 fn main() -> client::Result<()> {
@@ -16,33 +15,23 @@ fn main() -> client::Result<()> {
     // Setup the payer account
     let payer = setup_payer(&client, &config)?;
 
-    // Deploy the program
+    // Deploy or use existing program
     let program_id = setup_program(&client, &payer, &config)?;
 
     // Setup scheduler account
     let scheduler_account = setup_scheduler_account(&client, &payer, &program_id, &config)?;
 
     // Initialize the scheduler
-    initialize_scheduler(&client, &payer, &program_id, &scheduler_account, &config)?;
+    initialize_scheduler(&client, &payer, &program_id, &scheduler_account)?;
 
     // Create an Add task
     let add_task = Add::new(42, 58);
 
-    // Serialize the task
-    let task_data = serde_json::to_vec(&add_task).expect("Failed to serialize task");
-
     // Push the task onto the scheduler
-    push_task(
-        &client,
-        &payer,
-        &program_id,
-        &scheduler_account,
-        &task_data,
-        &config,
-    )?;
+    push_task(&client, &payer, &program_id, &scheduler_account, &add_task)?;
 
     // Execute the task
-    execute_task(&client, &payer, &program_id, &scheduler_account, &config)?;
+    execute_task(&client, &payer, &program_id, &scheduler_account)?;
 
     println!("Example completed successfully!");
     println!("The result of 42 + 58 has been computed and stored in the scheduler's data stack.");
